requirements.txt
Django
djangorestframework
djangorestframework_simplejwt
psycopg2-binary
python-dotenv
gunicorn
requests